refactor: relocate 4 endpoints to their perspective namespaces (D-017)
Breaking API change (clean break — the in-repo React SPA is the only consumer,
updated in the same commit). Moves four endpoints off the /api/scans/ namespace
to match the D-017 perspective model:

  GET  /api/scans/findings/             -> GET  /api/findings/
  POST /api/scans/findings/{id}/status/ -> POST /api/findings/{id}/status/
  GET  /api/scans/urls/                -> GET  /api/assets/urls/
  GET  /api/scans/deltas/              -> GET  /api/changes/

- findings_router added in findings/api.py (raw per-scan Finding instances,
  distinct from the /api/issues/ register); /urls/ added to the assets router
  (ordered before /{asset_id:int}/ so the int converter doesn't 422 on 'urls');
  deltas moved onto a new changes_router. Small serializers duplicated into the
  destination modules to avoid cross-layer import cycles (consistent with the
  per-module serializer style). scan_detail keeps its own copies.
- Registered /findings + /changes in ninja.py; removed the old scans routes.
- Frontend: FindingDetailModal -> /findings/{id}/status/, ChangesPage -> /changes/.
- Tests + CLAUDE.md URL layout updated.

~7 endpoints remain genuinely scan-centric under /api/scans/.
